SysInfo: Removed some warnings
git-svn-id: svn://ultimatepp.org/upp/trunk@13614 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ bazaar/SysInfo/MouseKey.cpp
@@ -223,7 +223,7 @@ bool SetKeyLockStatus(bool caps, bool num, bool scroll) {
 
 #ifdef PLATFORM_POSIX
 
-bool Mouse_GetPos(long &x, long &y) {
+bool Mouse_GetPos(int &x, int &y) {
 	SetSysInfoX11ErrorHandler();
 	_XDisplay *dpy = XOpenDisplay (NULL);
 	if (!dpy) {
